refactor: complete wave3 architecture decomposition (aud-h6b/h6c) (#595)
