+++ rulebook/tasks/issue-595-wave3-audit-remediation-architecture-decomposition/.metadata.json
@@ -1,0 +1,5 @@
+{
+  "status": "in_progress",
+  "createdAt": "2026-02-16T04:40:49.694Z",
+  "updatedAt": "2026-02-16T13:06:00.000Z"
+}
